# src/main/resources/openApi/openApi.yaml
openapi: 3.0.3
info:
  title: Standard API
  version: 1.0.0
servers:
  - url: https://localhost:8080/private/api/v1
...
paths:
...
  /private/api/v1/paymentMethod:
    post:
      summary: Save a new paymentMethod
      description: save and returns a new paymentMethod if successful.
      tags:
        - PaymentMethodController
      operationId: savePaymentMethod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: Successful save a new paymentMethod
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '400':
          description: Bad request (e.g., missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: find all brands
      description: return all the brands
      operationId: findAllPaymentMethods
      tags:
        - PaymentMethodController
      responses:
        '200':
          description: return all the brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethod'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/paymentMethod/{id}:
    get:
      summary: find by id
      description: return paymentMethod if is exist
      operationId: findPaymentMethodById
      tags:
        - PaymentMethodController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID PaymentMethod
      responses:
        '200':
          description: return a brands if exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: delete paymentMethod by id
      description: return status 200
      operationId: deletePaymentMethodById
      tags:
        - PaymentMethodController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID PaymentMethod
      responses:
        '200':
          description: delete has been successfully executed
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: return updated paymentMethod
      description: return updated paymentMethod
      operationId: updatePaymentMethod
      tags:
        - PaymentMethodController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID PaymentMethod
      requestBody:
        description: PaymentMethod to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: update has been successfully executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethod'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/provider:
    post:
      summary: Save a new provider
      description: save and returns a new provider if successful.
      tags:
        - ProviderController
      operationId: saveProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Successful save a new provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Bad request (e.g., missing or invalid fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: find all brands
      description: return all the brands
      operationId: findAllProviders
      tags:
        - ProviderController
      responses:
        '200':
          description: return all the brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/api/v1/provider/{id}:
    get:
      summary: find by id
      description: return provider if is exist
      operationId: findProviderById
      tags:
        - ProviderController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Provider
      responses:
        '200':
          description: return a brands if exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: delete provider by id
      description: return status 200
      operationId: deleteProviderById
      tags:
        - ProviderController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Provider
      responses:
        '200':
          description: delete has been successfully executed
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: return updated provider
      description: return updated provider
      operationId: updateProvider
      tags:
        - ProviderController
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: abc123
          description: The ID Provider
      requestBody:
        description: Provider to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: update has been successfully executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Bad request (e.g., invalid role)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (e.g., invalid role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Provider:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
        data:
          type: string
          format: date
        hora:
          type: string
          format: date-time
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        status:
          type: string
        description:
          type: string
        discountPercent:
          type: integer
        creationDate:
          type: string
          format: date-time
        creationTime:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        target:
          type: string
        message:
          type: string
        reference:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetails"
    ErrorDetails:
      type: object
      properties:
        target:
          type: string
        code:
          type: string
        message:
          type: string